UserRepository moved to Repository::User
Similar mass grep-and-replace as with Repository::Post.

RSpec: 864 examples, 0 failures, 1 pending
RuboCop: 274 files inspected, no offenses detected
