+++ spec/repositories/user_spec.rb
@@ -8,6 +8,8 @@ require_relative 'shared_examples/the_delete_method_for_a_repository'
 require_relative 'shared_examples/the_find_by_slug_method_for_a_repository'
 require_relative 'shared_examples/the_update_method_for_a_repository'
 
+require_relative '../../app/repositories/user'
+
 shared_examples 'a result with a Guest User entity' do
   it 'is a successful result with a Guest User entity' do
     expect(result).to be_success
@@ -16,7 +18,7 @@ shared_examples 'a result with a Guest User entity' do
   end
 end
 
-describe UserRepository do
+describe Repository::User do
   let(:be_entity_for) do
     ->(entity) { be_saved_user_entity_for(entity) }
   end
@@ -92,13 +94,13 @@ describe UserRepository do
     context 'for a dao with no passwords' do
       it 'adds test password' do
         dao = FactoryGirl.create :user
-        attr = UserRepository.new.send(:attributes_for, dao, [])
+        attr = Repository::User.new.send(:attributes_for, dao, [])
         expect(attr[:password]).to eq 'password'
       end
 
       it 'adds no test password per instruction' do
         dao = FactoryGirl.create :user
-        attr = UserRepository.new.send(:attributes_for, dao, [:no_password])
+        attr = Repository::User.new.send(:attributes_for, dao, [:no_password])
         expect(attr[:password]).to be_nil
       end
     end
@@ -140,4 +142,4 @@ describe UserRepository do
       end # context 'with a correct password'
     end # context 'for an existing user'
   end # describe :authenticate
-end # describe UserRepository
+end # describe Repository::User
